+++ storage/memory.ts
@@ -3,7 +3,7 @@ import { slugify } from "../core/utils/string.ts";
 import { contentType, globToRegExp, posix } from "../deps/std.ts";
 import { fromFilename } from "./transformers/mod.ts";
 
-import type { Data, Entry, EntryMetadata, Storage } from "../types.ts";
+import type { Data, Entry, EntrySource, Storage } from "../types.ts";
 
 export interface Options {
   root?: string;
@@ -54,7 +54,7 @@ export default class Memory implements Storage {
     }
   }
 
-  async *[Symbol.asyncIterator]() {
+  async *[Symbol.asyncIterator](): AsyncGenerator<EntrySource> {
     const root = this.root;
     const regexp = globToRegExp(posix.resolve(root, this.path));
 
@@ -65,11 +65,19 @@ export default class Memory implements Storage {
       const src = normalizePath(path);
       const name = src.slice(root.length + 1);
       yield {
-        label: name,
         name,
+        path: name,
         src,
       };
     }
+  }
+
+  source(name: string): EntrySource {
+    return {
+      src: posix.join(this.root, name),
+      name,
+      path: name,
+    };
   }
 
   name(name: string): string {
@@ -88,11 +96,7 @@ export default class Memory implements Storage {
   }
 
   get(name: string): Entry {
-    return new MemoryEntry({
-      src: posix.join(this.root, name),
-      name,
-      label: name,
-    }, this.#storage);
+    return new MemoryEntry(this.source(name), this.#storage);
   }
 
   delete(name: string) {
@@ -116,20 +120,16 @@ export default class Memory implements Storage {
 }
 
 export class MemoryEntry implements Entry {
-  metadata: EntryMetadata;
+  source: EntrySource;
   #storage: MemoryStorage;
 
-  constructor(metadata: EntryMetadata, storage: MemoryStorage) {
-    this.metadata = metadata;
+  constructor(source: EntrySource, storage: MemoryStorage) {
+    this.source = source;
     this.#storage = storage;
   }
 
-  get src(): string {
-    return this.metadata.src;
-  }
-
   readText(): Promise<string> {
-    const { src } = this.metadata;
+    const { src } = this.source;
     const content = this.#storage.get(src);
 
     if (content === undefined) {
@@ -144,14 +144,14 @@ export class MemoryEntry implements Entry {
   }
 
   writeText(content: string): Promise<void> {
-    const { src } = this.metadata;
+    const { src } = this.source;
     this.#storage.set(src, content);
 
     return Promise.resolve();
   }
 
   async readData(): Promise<Data> {
-    const { src } = this.metadata;
+    const { src } = this.source;
     const content = await this.readText();
     const transformer = fromFilename(src);
 
@@ -159,7 +159,7 @@ export class MemoryEntry implements Entry {
   }
 
   async writeData(data: Data) {
-    const { src } = this.metadata;
+    const { src } = this.source;
     const transformer = fromFilename(src);
     const content = (await transformer.fromData(data))
       .replaceAll(/\r\n/g, "\n"); // Unify line endings
@@ -168,7 +168,7 @@ export class MemoryEntry implements Entry {
   }
 
   readFile(): Promise<File> {
-    const { src, name } = this.metadata;
+    const { src, name } = this.source;
     const content = this.#storage.get(src);
     if (content === undefined) {
       throw new Error(`File not found: ${src}`);
@@ -182,7 +182,7 @@ export class MemoryEntry implements Entry {
   }
 
   async writeFile(file: File) {
-    const { src } = this.metadata;
+    const { src } = this.source;
     const content = await file.arrayBuffer();
     this.#storage.set(src, new Uint8Array(content));
   }